+++ src/api/admin/shipping-profiles/route.ts
@@ -1,5 +1,5 @@
 import { AuthenticatedMedusaRequest, MedusaResponse } from "@medusajs/framework/http"
-import { HttpTypes, UserDTO } from "@medusajs/framework/types"
+import { HttpTypes, StoreDTO } from "@medusajs/framework/types"
 import { refetchShippingProfile } from "@medusajs/medusa/api/admin/shipping-profiles/helpers"
 import { createShippingProfilesForStoreWorkflow } from "../../../workflows/create-shipping-profiles-for-store"
 
@@ -7,11 +7,11 @@ export const POST = async (
     req: AuthenticatedMedusaRequest<HttpTypes.AdminCreateShippingProfile>,
     res: MedusaResponse<HttpTypes.AdminShippingProfileResponse>
 ) => {
-    const loggedInUser = req.scope.resolve("loggedInUser") as UserDTO;
+    const currentStore = req.scope.resolve("currentStore") as Pick<StoreDTO, 'id'>;
     const shippingProfilePayload = req.validatedBody
 
     const { result } = await createShippingProfilesForStoreWorkflow(req.scope).run({
-        input: { data: [shippingProfilePayload], userId: loggedInUser.id },
+        input: { data: [shippingProfilePayload], storeId: currentStore.id },
     })
 
     const shippingProfile = await refetchShippingProfile(