+++ src/api/admin/stock-locations/middlewares.ts
@@ -4,10 +4,6 @@ import {
 } from "@medusajs/framework/http";
 import { addStoreIdToFilterableFields } from "../../middlewares/add-store-id-to-filterable-fields";
 import { moveIdsToQueryFromFilterableFields } from "../../middlewares/move-ids-to-query-from-filterable-fields";
-import {
-  UserType,
-  blockDataForUser,
-} from "../../middlewares/block-data-for-user";
 
 export const adminStockLocationRoutesMiddlewares: MiddlewareRoute[] = [
   {
@@ -20,7 +16,6 @@ export const adminStockLocationRoutesMiddlewares: MiddlewareRoute[] = [
         resourceId: "stock_location_id",
         filterableField: "store_id",
       }),
-      blockDataForUser(UserType.ADMIN),
       moveIdsToQueryFromFilterableFields,
     ],
   },