+++ src/api/middlewares/add-store-scope.ts
@@ -1,7 +1,7 @@
 import { type MedusaNextFunction, type MedusaRequest, type MedusaResponse } from "@medusajs/framework/http";
 import { ContainerRegistrationKeys } from "@medusajs/framework/utils";
 import { UserDTO } from "@medusajs/framework/types";
-import { SUPER_ADMIN_STORE_NAME } from "../../constants";
+import { container } from "@medusajs/framework"
 
 import * as cookie from "cookie";
 import { asValue } from "awilix";
@@ -32,6 +32,10 @@ export async function addStoreScope(req: MedusaRequest, res: MedusaResponse, nex
     req.scope.register({
       currentStore: asValue({ id }),
     });
+    // We use it for the import [product created hook], because we pause the workflow until it is confirmed. After confirmation, we get the container from the initial Medusa — https://github.com/medusajs/medusa/blob/develop/packages/modules/workflow-engine-inmemory/src/services/workflow-orchestrator.ts#L502
+    container.register({
+      currentStore: asValue({ id }),
+    });
   } else if (loggedInUser.metadata?.is_super_admin) {
     const { data: stores } = await query.graph({
       entity: "store",
@@ -45,6 +49,10 @@ export async function addStoreScope(req: MedusaRequest, res: MedusaResponse, nex
       },
     });
     req.scope.register({
+      currentStore: asValue({ id: stores[0].id }),
+    });
+    // We use it for the import [product created hook], because we pause the workflow until it is confirmed. After confirmation, we get the container from the initial Medusa — https://github.com/medusajs/medusa/blob/develop/packages/modules/workflow-engine-inmemory/src/services/workflow-orchestrator.ts#L502
+    container.register({
       currentStore: asValue({ id: stores[0].id }),
     });
   }