+++ src/workflows/link-customer-to-store/steps/get-current-store.ts
@@ -2,6 +2,6 @@ import { StoreDTO } from "@medusajs/framework/types";
 import { createStep, StepResponse } from "@medusajs/framework/workflows-sdk";
 
 export const getCurrentStoreStep = createStep("get-current-store-step", async ({}, { container }) => {
-  const currentStore = container.resolve("currentStore") as StoreDTO;
+  const currentStore = container.resolve("currentStore") as Pick<StoreDTO, 'id'>;
   return new StepResponse(currentStore);
 });